+++ pkg/test/auth/tokenmanager.go
@@ -1,4 +1,4 @@
-package test
+package auth
 
 import (
 	"crypto/rand"
