+++ pkg/test/auth/tokenmanager_test.go
@@ -1,4 +1,4 @@
-package test
+package auth
 
 import (
 	"bytes"
